Conditional aliases and installs

+++ config.sh
@@ -9,31 +9,41 @@
 # 4. Adds user to docker group
 # 5. Links files in dotfiles to home directory by relative path
 
-# apt installs
-sudo add-apt-repository ppa:libreoffice/ppa
-sudo apt update -y > /dev/null
-sudo apt install $(cat ./install_scripts/apt_installs_pop.txt) -y > /dev/null
-sudo npm install saltthepass > /dev/null
+# # apt installs
+if ! sudo -n true 2>/dev/null; then
+    echo "Please rerun as sudo to install packages"
+else
+    # Apt installs
+    sudo add-apt-repository ppa:libreoffice/ppa
+    sudo apt update -y > /dev/null
+    sudo apt install $(cat ./install_scripts/apt_installs_pop.txt) -y > /dev/null
+    sudo npm install saltthepass > /dev/null
 
-# npm
-while IFS= read -r package || [[ -n "$package" ]]; do
-    if [[ ! -z "$package" ]]; then
-        npm install -g "$package" > /dev/null
-    fi
-done < ./install_scripts/npm_installs_pop.txt
+    # npm
+    while IFS= read -r package || [[ -n "$package" ]]; do
+        if [[ ! -z "$package" ]]; then
+            npm install -g "$package" > /dev/null
+        fi
+    done < ./install_scripts/npm_installs_pop.txt
+
+    # DOCKER
+    sudo groupadd docker > /dev/null 2>&1
+    sudo usermod -aG docker $USER
+    sudo chmod 777 /var/run/docker.sock
+fi
+
 
 # Python installs
-pip install --upgrade pip > /dev/null 2>&1
-pip install -r ./install_scripts/py_installs_pop.txt > /dev/null 2>&1
+if command -v pip &> /dev/null; then
+    pip install --upgrade pip > /dev/null 2>&1
+    pip install -r ./install_scripts/py_installs_pop.txt > /dev/null 2>&1
+else
+    echo "Skipping python installs: pip not found"
+fi
 
-# DOCKER
-sudo groupadd docker > /dev/null 2>&1
-sudo usermod -aG docker $USER
-sudo chmod 777 /var/run/docker.sock
-
-# link dotfiles
+# # link dotfiles
 cd ~/dotfiles || { echo "Error: Could not access ~/dotfiles"; exit 1; }
-exclude=("install_scripts*" "*key" "code*" "READ*" "*osx" "*config.sh")
+exclude=("install_scripts*" "*key" "code*" "READ*" "*osx*" "*config.sh")
 git ls-files --exclude-standard | while read file; do # ignore .gitignore files
     for pattern in "${exclude[@]}"; do
         if [[ "$file" == $pattern ]]; then
@@ -42,14 +52,25 @@ git ls-files --exclude-standard | while read file; do # ignore .gitignore files
     done
     target="$HOME/$file"  # Place symlink in home directory
     mkdir -p "$(dirname "$target")"  # Ensure parent directories exist
-    rm "$target" || true
-    ln -s "$PWD/$file" "$target"
-    # echo "Linked $PWD/$file → $target"
+    if ! [ -f "$target" ]; then
+        echo "ADD  $PWD/$file → $target"
+        ln -s "$PWD/$file" "$target"
+    elif [[ "$file" == *git* || "$file" == *gh* ]]; then
+        echo "SKIP $PWD/$file → $target"
+    else
+        echo "REPL $PWD/$file → $target"
+        rm "$target" || true
+        ln -s "$PWD/$file" "$target"
+    fi
 done
 
-
-# dircolors, background
-eval `dircolors /home/$USER/.dir_colors/dircolors`
-gsettings set org.gnome.desktop.background picture-options 'none'
-gsettings set org.gnome.desktop.background primary-color '#282a36'
+HOST_FULL=$(hostname)
+HOST_PART=${HOST_FULL:0:3}
+if [[ $HOST_PART == "pop" ]]; then
+    echo "Pop!_OS detected"
+    dircolors, background
+    eval `dircolors /home/$USER/.dir_colors/dircolors`
+    gsettings set org.gnome.desktop.background picture-options 'none'
+    gsettings set org.gnome.desktop.background primary-color '#282a36'
+fi
